[radoub] test: Migrate seam-overlap + bone-transform + body-part-name tests to Radoub.UI.Tests (#2159)
Move tests for the helpers that migrated to MdlPartComposer / MdlPartNaming:
- ModelServiceSeamOverlapTests → MdlPartComposerSeamOverlapTests (6 tests)
- ModelServiceBoneTransformTests → MdlPartComposerBoneTransformTests (9 tests)
- ModelNameConstructionTests.BuildBodyPartName_* → MdlPartNamingTests (7 tests)

ModelNameConstructionTests retains BuildModelPrefix tests since BuildModelPrefix
(creature gender/race/phenotype) stays in QM.

All 32 migrated tests pass. QM ModelNameConstructionTests still passes (9/9).
