+++ Quartermaster/Quartermaster.Tests/ModelNameConstructionTests.cs
@@ -4,8 +4,9 @@ using Xunit;
 namespace Quartermaster.Tests;
 
 /// <summary>
-/// Tests for model name construction helpers in ModelService.
-/// Fixture-free — tests pure string formatting logic.
+/// Tests for QM-specific creature-prefix construction (gender/race/phenotype → "pmh0", "pfe2", etc.).
+/// The generic BuildBodyPartName helper moved to Radoub.UI in PR3a (#2159) — see
+/// MdlPartNamingTests in Radoub.UI.Tests.
 /// </summary>
 public class ModelNameConstructionTests
 {
@@ -28,7 +29,6 @@ public class ModelNameConstructionTests
         var result = ModelService.BuildModelPrefix(0, "H", 0);
         Assert.Equal("pmh0", result);
 
-        // Already lowercase
         var result2 = ModelService.BuildModelPrefix(0, "h", 0);
         Assert.Equal("pmh0", result2);
     }
@@ -45,28 +45,5 @@ public class ModelNameConstructionTests
     {
         var result = ModelService.BuildModelPrefix(1, "h", 0);
         Assert.StartsWith("pf", result);
-    }
-
-    [Theory]
-    [InlineData("pfo0", "head", 1, "pfo0_head001")]
-    [InlineData("pmh0", "chest", 1, "pmh0_chest001")]
-    [InlineData("pmd0", "neck", 2, "pmd0_neck002")]
-    [InlineData("pfe0", "footl", 1, "pfe0_footl001")]
-    [InlineData("pmh0", "bicepl", 10, "pmh0_bicepl010")]
-    [InlineData("pmh0", "shinr", 100, "pmh0_shinr100")]
-    public void BuildBodyPartName_FormatsCorrectly(string prefix, string partType, byte partNumber, string expected)
-    {
-        var result = ModelService.BuildBodyPartName(prefix, partType, partNumber);
-        Assert.Equal(expected, result);
-    }
-
-    [Fact]
-    public void BuildBodyPartName_PartNumberPaddedToThreeDigits()
-    {
-        var result = ModelService.BuildBodyPartName("pmh0", "head", 1);
-        Assert.EndsWith("001", result);
-
-        var result2 = ModelService.BuildBodyPartName("pmh0", "head", 99);
-        Assert.EndsWith("099", result2);
     }
 }
