+++ Radoub.UI/Radoub.UI.Tests/Services/MdlPartComposerSeamOverlapTests.cs
@@ -1,22 +1,23 @@
 using System.Numerics;
-using Quartermaster.Services;
 using Radoub.Formats.Mdl;
+using Radoub.UI.Services;
+using Xunit;
 
-namespace Quartermaster.Tests;
+namespace Radoub.UI.Tests.Services;
 
 /// <summary>
-/// Tests for ModelService.AdjustSeamOverlaps — validates that body parts with thin
-/// vertex overlap at joints get nudged closer together (#1557).
-/// Threshold is 0.10 world units (matching human-level overlap).
+/// Tests for MdlPartComposer.AdjustSeamOverlaps — body parts with thin vertex overlap
+/// at joints get nudged closer together (#1557). Threshold is 0.10 world units (matching
+/// human-level overlap). Originally added in Quartermaster; migrated when the seam-overlap
+/// helper moved into MdlPartComposer (PR3a, #2159).
 /// </summary>
-public class ModelServiceSeamOverlapTests
+public class MdlPartComposerSeamOverlapTests
 {
     private const float Tolerance = 0.001f;
-    private const float Threshold = 0.10f; // Must match ModelService.AdjustSeamOverlaps
+    private const float Threshold = 0.10f; // Must match MdlPartComposer.MinSeamOverlap
 
     /// <summary>
     /// Create a synthetic composite model with head and neck meshes at given Z positions.
-    /// Head mesh vertices span [headMinZ, headMaxZ], neck vertices span [neckMinZ, neckMaxZ].
     /// </summary>
     private static (MdlModel model, Dictionary<string, string> partTypes) CreateHeadNeckModel(
         float headMinZ, float headMaxZ, Vector3 headPosition,
@@ -76,20 +77,18 @@ public class ModelServiceSeamOverlapTests
     [Fact]
     public void AdjustSeamOverlaps_ThinOverlap_NudgesPartsCloser()
     {
-        // Head bottom at Z=1.0, neck top at Z=1.05 → overlap = 0.05 (below 0.10 threshold)
         var (model, partTypes) = CreateHeadNeckModel(
             headMinZ: 1.0f, headMaxZ: 1.5f, headPosition: new Vector3(0, 0, 1.25f),
             neckMinZ: 0.5f, neckMaxZ: 1.05f, neckPosition: new Vector3(0, 0, 0.75f));
 
-        float overlapBefore = 1.05f - 1.0f; // 0.05
+        float overlapBefore = 1.05f - 1.0f;
         Assert.True(overlapBefore < Threshold, "Precondition: overlap should be below threshold");
 
-        ModelService.AdjustSeamOverlaps(model, partTypes);
+        MdlPartComposer.AdjustSeamOverlaps(model, partTypes);
 
         var headMesh = model.GetMeshNodes().First(m => m.Name == "headmesh01");
         var neckMesh = model.GetMeshNodes().First(m => m.Name == "neckmesh01");
 
-        // Head moved down (Z decreased), neck moved up (Z increased)
         Assert.True(headMesh.Position.Z < 1.25f, "Head should have moved down");
         Assert.True(neckMesh.Position.Z > 0.75f, "Neck should have moved up");
     }
@@ -97,19 +96,17 @@ public class ModelServiceSeamOverlapTests
     [Fact]
     public void AdjustSeamOverlaps_AdequateOverlap_NoChange()
     {
-        // Head bottom at Z=1.0, neck top at Z=1.12 → overlap = 0.12 (above 0.10 threshold)
         var headPos = new Vector3(0, 0, 1.25f);
         var neckPos = new Vector3(0, 0, 0.75f);
         var (model, partTypes) = CreateHeadNeckModel(
             headMinZ: 1.0f, headMaxZ: 1.5f, headPosition: headPos,
             neckMinZ: 0.5f, neckMaxZ: 1.12f, neckPosition: neckPos);
 
-        ModelService.AdjustSeamOverlaps(model, partTypes);
+        MdlPartComposer.AdjustSeamOverlaps(model, partTypes);
 
         var headMesh = model.GetMeshNodes().First(m => m.Name == "headmesh01");
         var neckMesh = model.GetMeshNodes().First(m => m.Name == "neckmesh01");
 
-        // Positions should be unchanged
         Assert.Equal(headPos.Z, headMesh.Position.Z, Tolerance);
         Assert.Equal(neckPos.Z, neckMesh.Position.Z, Tolerance);
     }
@@ -117,37 +114,33 @@ public class ModelServiceSeamOverlapTests
     [Fact]
     public void AdjustSeamOverlaps_NegativeOverlap_NudgesPartsCloser()
     {
-        // Head bottom at Z=1.0, neck top at Z=0.99 → overlap = -0.01 (gap!)
         var (model, partTypes) = CreateHeadNeckModel(
             headMinZ: 1.0f, headMaxZ: 1.5f, headPosition: new Vector3(0, 0, 1.25f),
             neckMinZ: 0.5f, neckMaxZ: 0.99f, neckPosition: new Vector3(0, 0, 0.75f));
 
-        float overlapBefore = 0.99f - 1.0f; // -0.01 (gap)
+        float overlapBefore = 0.99f - 1.0f;
         Assert.True(overlapBefore < 0f, "Precondition: should be a gap");
 
-        ModelService.AdjustSeamOverlaps(model, partTypes);
+        MdlPartComposer.AdjustSeamOverlaps(model, partTypes);
 
         var headMesh = model.GetMeshNodes().First(m => m.Name == "headmesh01");
         var neckMesh = model.GetMeshNodes().First(m => m.Name == "neckmesh01");
 
-        // Both should have been nudged
         Assert.True(headMesh.Position.Z < 1.25f, "Head should have moved down");
         Assert.True(neckMesh.Position.Z > 0.75f, "Neck should have moved up");
 
-        // Total nudge should close the 0.01 gap plus reach the 0.10 threshold
         float totalNudge = (1.25f - headMesh.Position.Z) + (neckMesh.Position.Z - 0.75f);
-        Assert.Equal(0.11f, totalNudge, Tolerance); // deficit = 0.10 - (-0.01) = 0.11
+        Assert.Equal(0.11f, totalNudge, Tolerance);
     }
 
     [Fact]
     public void AdjustSeamOverlaps_SplitsDeficitEvenly()
     {
-        // Overlap = 0.05, threshold = 0.10, deficit = 0.05, half = 0.025
         var (model, partTypes) = CreateHeadNeckModel(
             headMinZ: 1.0f, headMaxZ: 1.5f, headPosition: new Vector3(0, 0, 1.25f),
             neckMinZ: 0.5f, neckMaxZ: 1.05f, neckPosition: new Vector3(0, 0, 0.75f));
 
-        ModelService.AdjustSeamOverlaps(model, partTypes);
+        MdlPartComposer.AdjustSeamOverlaps(model, partTypes);
 
         var headMesh = model.GetMeshNodes().First(m => m.Name == "headmesh01");
         var neckMesh = model.GetMeshNodes().First(m => m.Name == "neckmesh01");
@@ -155,15 +148,13 @@ public class ModelServiceSeamOverlapTests
         float headDelta = 1.25f - headMesh.Position.Z;
         float neckDelta = neckMesh.Position.Z - 0.75f;
 
-        // Both should move by the same amount
         Assert.Equal(headDelta, neckDelta, Tolerance);
-        Assert.Equal(0.025f, headDelta, Tolerance); // half of 0.05 deficit
+        Assert.Equal(0.025f, headDelta, Tolerance);
     }
 
     [Fact]
     public void AdjustSeamOverlaps_UnmappedMeshes_Ignored()
     {
-        // Mesh exists but isn't in the part type dictionary — should be skipped
         var root = new MdlNode { Name = "composite_root" };
         var mesh = new MdlTrimeshNode
         {
@@ -176,10 +167,9 @@ public class ModelServiceSeamOverlapTests
         root.Children.Add(mesh);
 
         var model = new MdlModel { Name = "test", GeometryRoot = root };
-        var partTypes = new Dictionary<string, string>(); // empty — nothing mapped
+        var partTypes = new Dictionary<string, string>();
 
-        // Should not throw, should not modify anything
-        ModelService.AdjustSeamOverlaps(model, partTypes);
+        MdlPartComposer.AdjustSeamOverlaps(model, partTypes);
 
         Assert.Equal(Vector3.Zero, mesh.Position);
     }
@@ -187,26 +177,23 @@ public class ModelServiceSeamOverlapTests
     [Fact]
     public void AdjustSeamOverlaps_ElfLikeOverlap_GetsNudged()
     {
-        // Simulate elf-like overlap: 0.048 (real measured value), below 0.10 threshold
+        // Real measured elf overlap: 0.048
         var (model, partTypes) = CreateHeadNeckModel(
             headMinZ: 1.655f, headMaxZ: 1.85f, headPosition: new Vector3(0, 0, 1.703f),
             neckMinZ: 1.50f, neckMaxZ: 1.703f, neckPosition: new Vector3(0, 0, 1.622f));
 
-        // Overlap = neckMaxZ - headMinZ = 1.703 - 1.655 = 0.048
         float overlapBefore = 1.703f - 1.655f;
         Assert.Equal(0.048f, overlapBefore, Tolerance);
         Assert.True(overlapBefore < Threshold);
 
-        ModelService.AdjustSeamOverlaps(model, partTypes);
+        MdlPartComposer.AdjustSeamOverlaps(model, partTypes);
 
         var headMesh = model.GetMeshNodes().First(m => m.Name == "headmesh01");
         var neckMesh = model.GetMeshNodes().First(m => m.Name == "neckmesh01");
 
-        // Head should move down, neck should move up
         Assert.True(headMesh.Position.Z < 1.703f, "Elf head should have moved down");
         Assert.True(neckMesh.Position.Z > 1.622f, "Elf neck should have moved up");
 
-        // Deficit = 0.10 - 0.048 = 0.052, each part moves 0.026
         float headDelta = 1.703f - headMesh.Position.Z;
         Assert.Equal(0.026f, headDelta, Tolerance);
     }
